+++ public/modules/lang/en.mjs
@@ -488,7 +488,7 @@ export const strings = {
     footPlayer:'Player', footCreate:'＋ Create a table', gameInfoTitle:'Game info', gameInfoEmpty:'Select a table',
     gfNone:'No game list filter', gfOpen:'Show open games', gfOpenNF:'Show open & non-full games', gfOpenNFNP:'Show open & non-full & non-private games', gfOpenNFP:'Show open & non-full & private games', gfOpenNFR:'Show open & non-full & ranking games',
     infoTypeLabel:'Type', infoCapitalLabel:'Capital', infoPlayersInGame:'Players in game', plIgnore:'Ignore', plStats:'Player stats',
-    plSortAlpha:'Sort alphabetically', plShowIdle:'Display idle players', plHideIdle:'Hide idle players',
+    plSortAlpha:'Sort alphabetically', plShowIdle:'Display idle players', plHideIdle:'Hide idle players', liveIdlePlayer:'Currently idle',
     plNotPlaying:'Not playing', plPlayingInFull:'%1 is playing in "%2".', plNotPlayingFull:'%1 is not playing at the moment.',
     plColumns:'Columns', plColAvatar:'Avatar', plColName:'Name', plColStatus:'In game', plColCountry:'Country', plColMe:'Me', plColActions:'Actions',
     gameTimeLabel:'Time', visRanked:'Ranked', infoBlindsUp:'Blinds up',